fixed forecast entity model and receiving data correctly now

+++ app/src/main/java/com/example/weatherapp5/weatherdetail/forecast/domain/entity/Main.kt
@@ -28,6 +28,6 @@ class Main {
     var humidity: Int? = null
     @SerializedName("temp_kf")
     @Expose
-    var tempKf: Int? = null
+    var tempKf: Double? = null
 
 }
